resolved error about navigation functionality in auth service
