+++ lib/main.dart
@@ -2,8 +2,8 @@ import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:pinterest_clone/firebase_options.dart';
-import 'package:pinterest_clone/screens/landing_page.dart';
-import 'package:pinterest_clone/services/auth_services.dart';
+
+import 'screens/pinterest_clone.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized(); 
@@ -20,7 +20,7 @@ class MainApp extends StatelessWidget {
   Widget build(BuildContext context) {
     return MaterialApp(
       debugShowCheckedModeBanner: false,
-      home: TestAuthPage()
+      home: PinterestClonePage()
     );
   }
 }
