+++ lib/providers/auth_providers.dart
@@ -1,0 +1,7 @@
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+
+import '../services/auth_services.dart';
+
+final authServicesProvider = Provider<AuthServices>((ref) {
+  return AuthServices();
+});
