+++ lib/screens/main-screens/home_page.dart
@@ -2,7 +2,7 @@ import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../../services/auth_services.dart';
+import '../../providers/auth_providers.dart';
 
 class HomePage extends ConsumerWidget {
   HomePage({super.key});
@@ -15,7 +15,7 @@ class HomePage extends ConsumerWidget {
     return Scaffold(
       appBar: AppBar(leading: GestureDetector( 
         onTap: (){
-            authProvider.signUserOut();
+            authProvider.signUserOut(context);
           },child: Icon(Icons.exit_to_app)
         ),
       ),
