+++ lib/screens/pinterest_clone.dart
@@ -1,0 +1,29 @@
+import 'package:firebase_auth/firebase_auth.dart';
+import 'package:flutter/material.dart';
+
+import 'landing_page.dart';
+import 'main-screens/home_page.dart';
+
+class PinterestClonePage extends StatelessWidget {
+  const PinterestClonePage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: StreamBuilder<User?>(
+        stream: FirebaseAuth.instance.authStateChanges(),
+        builder: ((context, snapshot) {
+          // print("Auth state changed: ${snapshot.connectionState}");
+          if (snapshot.hasData){
+            // print("User is logged in: ${snapshot.data?.email}");
+            return HomePage();
+          }
+          else{
+            // print("User is not logged in");
+            return const LandingPage();
+          }
+        }),
+      ),
+    );
+  }
+}
